feat: compress to multi, decompress single feat

+++ Decompressor.java
@@ -10,11 +10,18 @@ public class Decompressor {
     private String inputDir;
     private String outputDir;
     private int readBufferSize;
+    private String prefix;
+    private int fileNumber;
+    private String suffix;
 
-    public Decompressor(String inputDir, String outputDir, int readBufferSize) {
+    public Decompressor(String inputDir, String outputDir, int readBufferSize,
+                        String prefix, String suffix) {
         this.inputDir = inputDir;
         this.outputDir = outputDir;
         this.readBufferSize = readBufferSize;
+        this.prefix = prefix;
+        this.suffix = suffix;
+        this.fileNumber = 1;
     }
 
     public void decompress() {
@@ -22,39 +29,41 @@ public class Decompressor {
         if (!outFolder.exists())
             outFolder.mkdir();
 
-        String ZipFile = this.inputDir + "/" + "file.agozip";
-        try {
-            ZipInputStream zis = new ZipInputStream(new FileInputStream(ZipFile));
-            ZipEntry zipEntry = zis.getNextEntry();
-
-            byte[] buffer = new byte[this.readBufferSize];
-            while (zipEntry != null) {
-                File newFile = newFile(outFolder, zipEntry);
-                if (zipEntry.isDirectory()) {
-                    if (!newFile.isDirectory() && !newFile.mkdirs()) {
-                        throw new IOException("Failed to create directory " + newFile);
+        String zipFileStr = this.inputDir + "/" + this.prefix + this.fileNumber + this.suffix;
+        File zipFile = new File(zipFileStr);
+        if (zipFile.exists()) {
+            try {
+                ZipInputStream zis = new ZipInputStream(new FileInputStream(zipFileStr));
+                ZipEntry zipEntry = zis.getNextEntry();
+                byte[] buffer = new byte[this.readBufferSize];
+                while (zipEntry != null) {
+                    File newFile = newFile(outFolder, zipEntry);
+                    if (zipEntry.isDirectory()) {
+                        if (!newFile.isDirectory() && !newFile.mkdirs()) {
+                            throw new IOException("Failed to create directory " + newFile);
+                        }
+                    } else {
+                        File parent = newFile.getParentFile();
+                        if (!parent.isDirectory() && !parent.mkdirs()) {
+                            throw new IOException("Failed to create directory " + parent);
+                        }
+                        FileOutputStream fis = new FileOutputStream(newFile);
+                        int len;
+                        while ((len = zis.read(buffer)) > 0)
+                            fis.write(buffer, 0, len);
+                        fis.close();
                     }
-                } else {
-                    File parent = newFile.getParentFile();
-                    if (!parent.isDirectory() && !parent.mkdirs()) {
-                        throw new IOException("Failed to create directory " + parent);
-                    }
-                    FileOutputStream fis = new FileOutputStream(newFile);
-                    int len;
-                    while ((len = zis.read(buffer)) > 0)
-                        fis.write(buffer, 0, len);
-                    fis.close();
+                zipEntry = zis.getNextEntry();
                 }
-            zipEntry = zis.getNextEntry();
+                zis.closeEntry();
+                zis.close();
+            } catch (IOException e) {
+                System.out.println(e);
             }
-            zis.closeEntry();
-            zis.close();
-        } catch (IOException e) {
-            System.out.println(e);
         }
     }
 
-    private static File newFile(File destDir, ZipEntry zipEntry) throws IOException {
+    private File newFile(File destDir, ZipEntry zipEntry) throws IOException {
         File destFile = new File(destDir, zipEntry.getName());
         String destDirPath = destDir.getCanonicalPath();
         String destFilePath = destFile.getCanonicalPath();
